comparateur-devis/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.4.2
uuid==1.30
anthropic